+++ src/main/java/com/tictac/demo/service/ProcesoService.java
@@ -19,11 +19,36 @@ public class ProcesoService {
     }
 
     public Proceso createProceso(Proceso proceso){
-        return procesoRepository.save(proceso);
+        if(proceso.getIdMomento() == null || proceso.getIdMomento().toString().isEmpty() ||
+            proceso.getDescripcion() == null || proceso.getDescripcion().trim().isEmpty() ||
+            proceso.getTiempo() == null || proceso.getTiempo().trim().isEmpty()){
+            return null;
+        }else{
+            return procesoRepository.save(proceso);
+        }
     }
 
-    public void deleteProceso(Integer id){
-        procesoRepository.deleteById(id);
+    public String updateProceso(Proceso proceso){
+        if(procesoRepository.existsById(proceso.getIdProceso())){
+            Optional<Proceso> p = procesoRepository.findById(proceso.getIdProceso());
+
+            p.get().setIdMomento(proceso.getIdMomento());
+            p.get().setDescripcion(proceso.getDescripcion());
+            p.get().setTiempo(proceso.getTiempo());
+            procesoRepository.save(p.get());
+            return "Proceso actualizado con éxito";
+        }else{
+            return null;
+        }
+    }
+
+    public String deleteProceso(Integer id){
+        if(procesoRepository.existsById(id)){
+            procesoRepository.deleteById(id);
+            return "Proceso eliminado con éxito";
+        }else{
+          return null;
+        }
     }
 
     public List<Proceso> listProceso(){
