+++ src/main/java/com/tictac/demo/service/ProyectoAulaService.java
@@ -19,11 +19,44 @@ public class ProyectoAulaService {
     }
 
     public ProyectoAula createProyectoAula(ProyectoAula proyectoAula){
-        return proyectoAulaRepository.save(proyectoAula);
+        if(proyectoAula.getGrado() == null || proyectoAula.getGrado().toString().trim().isEmpty() ||
+            proyectoAula.getIdTema() == null || proyectoAula.getIdTema().toString().trim().isEmpty() ||
+            proyectoAula.getDocenteLider() == null || proyectoAula.getDocenteLider().trim().isEmpty() ||
+            proyectoAula.getFechaInicio() == null || proyectoAula.getFechaInicio().toString().trim().isEmpty() ||
+            proyectoAula.getFechaFin() == null || proyectoAula.getFechaFin().toString().trim().isEmpty() ||
+            proyectoAula.getLeccionesAprendidas() == null || proyectoAula.getLeccionesAprendidas().trim().isEmpty() ||
+            proyectoAula.getNombre() == null || proyectoAula.getNombre().trim().isEmpty()){
+            return null;
+        }else{
+            return proyectoAulaRepository.save(proyectoAula);
+        }
     }
 
-    public void deleteProyectoAula(Integer id){
-        proyectoAulaRepository.deleteById(id);
+    public String updateProyectoAula(ProyectoAula proyectoAula){
+        if(proyectoAulaRepository.existsById(proyectoAula.getIdProyecto())){
+            Optional<ProyectoAula> pa = proyectoAulaRepository.findById(proyectoAula.getIdProyecto());
+
+            pa.get().setGrado(proyectoAula.getGrado());
+            pa.get().setIdTema(proyectoAula.getIdTema());
+            pa.get().setDocenteLider(proyectoAula.getDocenteLider());
+            pa.get().setFechaInicio(proyectoAula.getFechaInicio());
+            pa.get().setFechaFin(proyectoAula.getFechaFin());
+            pa.get().setLeccionesAprendidas(proyectoAula.getLeccionesAprendidas());
+            pa.get().setNombre(proyectoAula.getNombre());
+            proyectoAulaRepository.save(pa.get());
+            return "Proyecto de aula actualizado con éxito";
+        }else{
+            return null;
+        }
+    }
+
+    public String deleteProyectoAula(Integer id){
+        if(proyectoAulaRepository.existsById(id)){
+            proyectoAulaRepository.deleteById(id);
+            return "Proyecto de aula eliminado con éxito";
+        }else{
+            return null;
+        }
     }
 
     public List<ProyectoAula> listProyectoAula(){
